test(sunburst): 暂时修复下旭日图单测
